Bugfixes, first successful run
