+++ baker_wet_cleaning_application/scripts/map_handling_behavior.py
@@ -55,7 +55,7 @@ class MapHandlingBehavior(behavior_container.BehaviorContainer):
 			)
 		self.room_sequencer_.executeCustomBehavior()
 		self.room_sequencing_data_ = self.room_sequencer_.room_sequence_result_	
-		self.database_handler_.setRoomMapping(self.rooms_list_, self.room_sequencing_data_)
+		self.mapping_ = self.database_handler_.getRoomMapping(self.rooms_list_, self.room_sequencing_data_)
 		#self.printMsg("self.room_sequencing_data_.checkpoints=" + str(self.room_sequencing_data_.checkpoints))
 		#self.printMsg("self.segmentation_data_.room_information_in_pixel=" + str(self.segmentation_data_.room_information_in_pixel))
 		#self.printMsg("self.segmentation_data_.room_information_in_meter=" + str(self.segmentation_data_.room_information_in_meter))
